Add option to display line numbers.
This currently prevents word wrapping while line numbers are displayed.
Getting around that limitation might require a more ambitious editor
replacement. I'd rather avoid doing that, since custom editors tend to
come with a variety of corner-case usability issues, but it might be
something to look at in the future.