Allow rendering delegation to helper extensions.
It is now possible to have multiple top-level context menu items,
via extra extensions that are solely responsible for displaying
the menu items and receiving the click events. All execution logic
continues to live within the main Contextlets extension.

When these helper extensions are installed and enabled, a new
drop-down field will exist in the main extension's menu item
configuration interface, which allows the user to select which
extension will be responsible for displaying the item.

Any extension whose name begins with "extra contextlet" (case-
insensitive) will be listed for selection in the configuration
interface.