Fix issue where deleted items sometimes don't cause an update
When deleting the last remaining item in a particular renderer,
the renderer would be skipped in the update push because it
didn't have any items. This meant the deleted item would continue
to show up in the context menu.

Now, the push will apply to all renderers that had items at any
point during the lifecycle of the main extension.